Adding new feature for customizing up- and downscripts using templates.
We could certainly handle this in a different way, or use a different
location for ucarp::params::scriptDir.  We could also expose
ucarp::params::scriptDir if people want to put their scripts in a different
location.

The real limiting factor here is the initscript configuration: in RHEL
at least, the scriptdir is configured inside the initscript and can't
be modified via configuration files.  You actually have to modify the
initscript itself.  Not a major problem to do so, but any time I can
avoid customizing initscripts, so much the better.